# api-docs.yaml
openapi: 3.0.1
info:
  title: car api
  description: Api for car entity
  version: "1.0"
servers:
- url: http://localhost:9005
  description: Generated server url
paths:
  /car/v1/edit:
    put:
      tags:
      - car-controller
      operationId: editCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /car/v1/add:
    post:
      tags:
      - car-controller
      operationId: addCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /car/v1/all:
    get:
      tags:
      - car-controller
      operationId: fetchAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Car'
  /car/v1/delete:
    delete:
      tags:
      - car-controller
      operationId: deleteCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    Car:
      required:
      - brand
      - carModel
      - category
      - mileage
      - price
      - releaseDate
      type: object
      properties:
        id:
          type: integer
          format: int32
        brand:
          type: string
        carModel:
          type: string
        category:
          type: string
        price:
          type: integer
          format: int64
        mileage:
          type: number
          format: float
        releaseDate:
          type: string